feat(config-editor): the Help page is a page, not a hosted dialog
Help.vue absorbs HelpModal's content and the component is deleted. The
dialog framing — a 60rem card centred in the page, modal chrome, its
own inner scroll region — becomes the same full-bleed layout as the
configuration editor page: viewport-height scroll container, matching
title and tab styles, and a sticky title/tab strip so the sections
stay reachable in a long panel.

The prose fills the width it gets: panels flow into 30rem CSS columns
(columns by width, not count, so the browser fits one, two or three to
the window with no media-query ladder), and break-inside keeps a
heading with its body. Keyboard/ARIA tab behaviour is unchanged.

Two comments elsewhere cited HelpModal as precedent; they now cite the
Help page.
